requirements.txt
discord.py
pynput
asyncio
python-dotenv
colorama